refactor: removes filterapi/x package
