+++ cmd/extproc/mainlib/main.go
@@ -29,7 +29,6 @@ import (
 	"google.golang.org/grpc/credentials/insecure"
 	"google.golang.org/grpc/health/grpc_health_v1"
 
-	"github.com/envoyproxy/ai-gateway/filterapi/x"
 	"github.com/envoyproxy/ai-gateway/internal/extproc"
 	"github.com/envoyproxy/ai-gateway/internal/internalapi"
 	"github.com/envoyproxy/ai-gateway/internal/metrics"
@@ -151,7 +150,7 @@ func Main(ctx context.Context, args []string, stderr io.Writer) (err error) {
 	}
 
 	metricsServer, meter := startMetricsServer(metricsLis, l)
-	chatCompletionMetrics := metrics.NewChatCompletion(meter, x.NewCustomChatCompletionMetrics, metricsRequestHeaderLabelMapping)
+	chatCompletionMetrics := metrics.NewChatCompletion(meter, metricsRequestHeaderLabelMapping)
 	embeddingsMetrics := metrics.NewEmbeddings(meter, metricsRequestHeaderLabelMapping)
 
 	server, err := extproc.NewServer(l)
