+++ internal/metrics/embeddings_metrics.go
@@ -11,7 +11,7 @@ import (
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/metric"
 
-	"github.com/envoyproxy/ai-gateway/filterapi/x"
+	"github.com/envoyproxy/ai-gateway/filterapi"
 )
 
 // embeddings is the implementation for the embeddings AI Gateway metrics.
@@ -19,8 +19,24 @@ type embeddings struct {
 	baseMetrics
 }
 
+// EmbeddingsMetrics is the interface for the embeddings AI Gateway metrics.
+type EmbeddingsMetrics interface {
+	// StartRequest initializes timing for a new request.
+	StartRequest(headers map[string]string)
+	// SetModel sets the model the request. This is usually called after parsing the request body .
+	SetModel(model string)
+	// SetBackend sets the selected backend when the routing decision has been made. This is usually called
+	// after parsing the request body to determine the model and invoke the routing logic.
+	SetBackend(backend *filterapi.Backend)
+
+	// RecordTokenUsage records token usage metrics for embeddings (only input and total tokens are relevant).
+	RecordTokenUsage(ctx context.Context, inputTokens, totalTokens uint32, requestHeaderLabelMapping map[string]string, extraAttrs ...attribute.KeyValue)
+	// RecordRequestCompletion records latency metrics for the entire request.
+	RecordRequestCompletion(ctx context.Context, success bool, requestHeaderLabelMapping map[string]string, extraAttrs ...attribute.KeyValue)
+}
+
 // NewEmbeddings creates a new Embeddings instance.
-func NewEmbeddings(meter metric.Meter, requestHeaderLabelMapping map[string]string) x.EmbeddingsMetrics {
+func NewEmbeddings(meter metric.Meter, requestHeaderLabelMapping map[string]string) EmbeddingsMetrics {
 	return &embeddings{
 		baseMetrics: newBaseMetrics(meter, genaiOperationEmbedding, requestHeaderLabelMapping),
 	}
