--- tests/extproc/vcr/extproc_custom_metrics.yaml
@@ -1,21 +1,0 @@
-# Copyright Envoy AI Gateway Authors
-# SPDX-License-Identifier: Apache-2.0
-# The full text of the Apache license is available in the LICENSE file at
-# the root of the repo.
-
-# Same as extproc.yaml, except including llmRequestCosts for dynamic metadata
-
-metadataNamespace: ai_gateway_llm_ns
-schema:
-  name: OpenAI
-  version: v1
-modelNameHeaderKey: x-model-name
-backends:
-  - name: openai
-    schema:
-      name: OpenAI
-      version: v1
-# Required for dynamic metadata to be built
-llmRequestCosts:
-  - metadataKey: "test_cost"
-    type: "InputToken"
